+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### 0.5.3
 
+- [+] 新增了 `useLogger` 的全局 props。`useLogger={true}` 时，每当用户填写表单时，在 console 里的展示类似如下：
+  <img src="https://img.alicdn.com/tfs/TB11rt_AbY1gK0jSZTEXXXDQVXa-1336-468.jpg" width="500" />
+  自下向上一层层展示用户触发的 formData 变化，便于开发者快速定位问题。
 - [+] 列表新增 `ui:options` 字段 `itemButtons`, 用于添加每个 item 的操作按钮（默认是一个删除按钮）。具体使用方式参见文档 ui-schema。
 - [!] 增加了列表的收缩按钮的点击面积